Number pairs components test:     -- minor package structure refactoring     -- tests written and passing
