+++ components-impl/src/main/java/ru/rougegibbons/landsanddungeons/components/models/core/SingleIntComponentModel.java
@@ -1,0 +1,29 @@
+package ru.rougegibbons.landsanddungeons.components.models.core;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.jetbrains.annotations.NotNull;
+import ru.rougegibbons.landsanddungeons.components.core.numbers.integer.SingleIntComponentImpl;
+import ru.rougegibbons.landsanddungeons.components.models.ComponentModel;
+
+/**
+ * data model for {@link SingleIntComponentImpl} class.
+ * @see SingleIntComponentImpl
+ */
+public class SingleIntComponentModel extends ComponentModel {
+    private final Integer value;
+
+    public SingleIntComponentModel(@NotNull @JsonProperty("id") Long id,
+                                   @NotNull @JsonProperty("value") Integer value) {
+        super(id);
+        this.value = value;
+    }
+
+    /**
+     * get stored data.
+     * @return stored data.
+     */
+    @JsonProperty("value")
+    public @NotNull Integer getValue() {
+        return value;
+    }
+}
