+++ components-impl/src/main/java/ru/rougegibbons/landsanddungeons/components/models/core/StringComponentModel.java
@@ -1,0 +1,34 @@
+package ru.rougegibbons.landsanddungeons.components.models.core;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.jetbrains.annotations.NotNull;
+import ru.rougegibbons.landsanddungeons.components.core.string.StringComponentImpl;
+import ru.rougegibbons.landsanddungeons.components.models.ComponentModel;
+
+/**
+ * data model class for {@link StringComponentImpl} class.
+ * @see StringComponentImpl
+ */
+public class StringComponentModel extends ComponentModel {
+    private final String text;
+
+    /**
+     * constructor for deserialization from resource file/package or for component instance serialization.
+     * @param id - instance id.
+     * @param text - stored text
+     */
+    public StringComponentModel(@NotNull @JsonProperty("id") Long id,
+                                @NotNull @JsonProperty("text") String text) {
+        super(id);
+        this.text = text;
+    }
+
+    /**
+     * get access to stored text.
+     * @return text, stored in model.
+     */
+    @JsonProperty("text")
+    public @NotNull String getText() {
+        return text;
+    }
+}
