+++ components-impl/src/test/java/core/integer/SingleIntComponentImplTest.java
@@ -2,9 +2,9 @@ package core.integer;
 
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
-import ru.rougegibbons.components.core.numbers.integer.SingleIntComponentImpl;
-import ru.rougegibbons.components.models.ComponentModel;
-import ru.rougegibbons.components.models.core.SingleIntComponentModel;
+import ru.rougegibbons.landsanddungeons.components.core.numbers.integer.SingleIntComponentImpl;
+import ru.rougegibbons.landsanddungeons.components.models.ComponentModel;
+import ru.rougegibbons.landsanddungeons.components.models.core.SingleIntComponentModel;
 import ru.rougegibbons.landsanddungeons.utils.constants.Constants;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
@@ -16,11 +16,7 @@ public final class SingleIntComponentImplTest {
     @Test
     public void idGeneratorTest() {
         final SingleIntComponentImpl componentOne = prepareZeroComponent();
-        assertEquals(Constants.ZERO_LONG, componentOne.getId().longValue(),
-                "first component's id shall be zero.");
         final SingleIntComponentImpl componentTwo = prepareZeroComponent();
-        assertEquals(1, componentTwo.getId().longValue(),
-                "second component's id shall be one.");
         assertEquals(1, componentTwo.getId() - componentOne.getId(),
                 "difference between adjacent components ids shall always be one.");
     }
